requirements.txt
kafka-python>=2.0.2
pyspark>=3.3.0
apache-airflow==2.6.3
psycopg2-binary>=2.9.6
pandas>=2.0.0
faker>=18.9.0
matplotlib>=3.7.0
seaborn>=0.12.2
